// frontend/src/api/generated/schemas/sessionDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * onBoard API
 * AI-powered interview preparation service
 * OpenAPI spec version: 1.0
 */
import type { SessionDtoStatus } from "./sessionDtoStatus";
import type { SessionTechnologyLevelDto } from "./sessionTechnologyLevelDto";

export interface SessionDto {
  id: string;
  userId: string;
  technologyLevelId: string;
  status: SessionDtoStatus;
  /** @nullable */
  totalQuestions: number | null;
  currentOrder: number;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  finishedAt: string | null;
  createdAt: string;
  technologyLevel: SessionTechnologyLevelDto;
}